TVB_requirements.txt
absl_py==2.0.0+computecanada
anytree==2.12.1
appdirs==1.4.4+computecanada
autopep8==2.0.4
backcall==0.2.0+computecanada
bctpy==0.6.1
bidict==0.22.1
bioread==3.0.1
biosppy==1.0.0
cached-property==1.5.2+computecanada
cachetools==5.3.2+computecanada
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.1.4+computecanada
cvxopt==1.2.6+computecanada
cycler==0.11.0+computecanada
dataclasses==0.6+computecanada
decorator==5.1.1+computecanada
docopt==0.6.2+computecanada
ebrains-drive==0.6.0
filelock==3.12.2+computecanada
future==0.18.2+computecanada
google-auth==2.25.0
google_auth_oauthlib==0.4.6+computecanada
grpcio==1.38.1+computecanada
h5py==3.4.0+computecanada
idna==3.6
imagecodecs==2020.2.18+computecanada
imageio==2.31.2
importlib-metadata==6.7.0
ipython==7.34.0+computecanada
ipywidgets==8.1.1+computecanada
jedi==0.19.1+computecanada
Jinja2==3.1.2+computecanada
joblib==1.3.2+computecanada
jupyterlab_widgets==3.0.9+computecanada
kiwisolver==1.3.1+computecanada
littleutils==0.2.2
llvmlite==0.36.0+computecanada
lxml==4.7.1+computecanada
Mako==1.2.4+computecanada
Markdown==3.4.4+computecanada
MarkupSafe==2.1.3
matplotlib==3.4.2+computecanada
matplotlib_inline==0.1.6+computecanada
mne==1.3.1
networkx==2.6.3+computecanada
neuroglancer-scripts==1.1.0
neurokit==0.2.0
nflows==0.14
nibabel==4.0.2
nilearn==0.10.2
nolds==0.5.2
numba==0.53.1+computecanada
numexpr==2.7.0+computecanada
numpy==1.20.0+computecanada
oauthlib==3.2.2+computecanada
opencv-python==4.5.1.48+computecanada
opt_einsum==3.3.0+computecanada
outdated==0.2.2
packaging==23.2+computecanada
pandas==1.3.5+computecanada
pandas-flavor==0.6.0
parso==0.8.3+computecanada
patsy==0.5.4
pexpect==4.9.0
pickleshare==0.7.5+computecanada
Pillow==8.4.0+computecanada
pingouin==0.5.3
pip==23.3.1
platformdirs==4.0.0
pooch==1.8.0+computecanada
prompt_toolkit==3.0.41+computecanada
protobuf==3.20.3
psutil==5.8.0+computecanada
ptyprocess==0.7.0+computecanada
pyasn1==0.5.1+computecanada
pyasn1_modules==0.3.0+computecanada
pycodestyle==2.10.0
pygments==2.17.2+computecanada
PyJWT==2.8.0
pyknos==0.15.2
PyLEMS==0.6.4
pyparsing==3.1.1+computecanada
pyro-ppl==1.3.1
pyro_api==0.1.2+computecanada
python-dateutil==2.8.2+computecanada
pytz==2023.3.post1+computecanada
pyunicore==0.9.12
PyWavelets==1.1.1+computecanada
requests==2.31.0+computecanada
requests_oauthlib==1.3.1+computecanada
rsa==4.9+computecanada
sbi==0.14.2
scikit-image==0.18.2+computecanada
scikit_learn==1.0.1+computecanada
scipy==1.7.3+computecanada
seaborn==0.12.2+computecanada
setuptools==68.0.0+computecanada
shortuuid==1.0.11
# Editable install with no version control (showcase1-ageing==0.1.0)
-e /home/${USER}/TVB/virtual_ageing
siibra==1.0a2
six==1.16.0+computecanada
sklearn==0.0+computecanada
-e git+https://github.com/ins-amu/virtual_aging_brain.git@32efae569bb7984b69ca67d3b1dd30d5a53aa58a#egg=src
statsmodels==0.13.5
tabulate==0.9.0+computecanada
tensorboard==2.11.2+computecanada
tensorboard-data-server==0.6.1+computecanada
tensorboard_plugin_wit==1.8.1+computecanada
threadpoolctl==3.1.0+computecanada
tifffile==2021.11.2+computecanada
tomli==2.0.1+computecanada
torch==1.6.0+computecanada
tqdm==4.66.1+computecanada
traitlets==5.9.0+computecanada
tvb-data==2.0
tvb-ebrains-data @ git+https://gitlab.ebrains.eu/fousekjan/tvb-ebrains-data.git@0.2.5
#git+https://gitlab.ebrains.eu/fousekjan/tvb-ebrains-data.git@f2a35c6bb68a4c4f14309c49e515f33b7d4aa2dd
tvb-library @ git+https://github.com/the-virtual-brain/tvb-root/@parameters-api#egg=tvb-library&subdirectory=scientific_library
#git+https://github.com/the-virtual-brain/tvb-root/@f570bf618fbb885ec12bf36ab920f3b43d0d8c27#subdirectory=scientific_library
typing==3.7.4.3+computecanada
typing_extensions==4.7.1+computecanada
urllib3==2.0.7
wcwidth==0.2.12+computecanada
Werkzeug==2.2.3+computecanada
wheel==0.42.0
widgetsnbextension==4.0.9+computecanada
xarray==0.20.2+computecanada
zipp==3.15.0+computecanada
